Back out the experimental spare bits calculation and everything that depended on it
The experimental spare bits calculation helped to validate
the structure of this code and the handling of various
multi-payload enum details, but it cannot be done here
in a way that will always agree with the compiler.

So I've backed that out:  The result can still handle
certain multi-payload enums (those with dynamic layouts that
can never utilize spare bits) but the remainder will require
some additional work for the compiler to expose the real
spare bit data for use in the runtime.