[SE-0474] Support `yielding borrow` and `yielding mutate` accessors
`yielding borrow` and `yielding mutate` are the final approved
names for the accessors previously known as `read` and `modify`
(which are themselves improvements over the earlier non-standard
`_read` and `_modify` accessors).

Since some people are already using these under the `read` and
`modify` names, we want to support the old and new names as synonyms
for a transition period until everyone has had a chance to migrate.